requirements.txt
# Core dependencies
mcp>=1.0.0
httpx>=0.27.0
pydantic>=2.0.0
python-dotenv>=1.0.0

# Development dependencies (optional)
pytest>=8.0.0
pytest-asyncio>=0.23.0
